Add basic options to each tab
Added CreateLayout(...) function and use it to create the layout for each tab. Added basic enabled settings to each tab. Removed the underscore key from the channel input.
Vertical layouts with certain properties set have to be used instead of scrollable containers because the latter has issues with sizing and being disabled. Rough draft with a bunch of issues. The mod MoreButtons adds an underscore and more symbols to the keyboard itself.