Add websocket tab to config menu
Added the two websocket settings to a new tab. General tab is shown by default instead of all of them. Widened tab selector.
Websocket settings are supposed to be on a separate panel, but an extra tab is quicker to get working for now. Tab selector widened to fit more tabs and create more clarity between the selector and tabs themselves.