Fix panelPositionButtonRow height bug
Fixed a bug where the HorizontalLayoutGroup for the menu / game position buttons in the panel tab wasn't calculating its height until hovered over or the tab was disabled and re-enabled by adding a 0.15 second delay to hiding tabs.
Buttons intersecting with the tab selector and general weird behavior looks very weird and broken.